requirements.txt
kafka-python
PyGithub
